requirements.txt
woocommerce
requests
openai
langchain
langchain-core
langchain-community
langchain-openai
langchain_experimental
python-dotenv